Change grid and container sizes to px
Changes grid and container sizes to `px`, as the
viewport pixel size does not depend on the font size.

The actual em values were inconsistent with the docs,
while the docs were also inconsistent with the comments, namely:

* `sm` breakpoint was 34em which is 558px, not 480px.
  Changed to 480px, matching the docs and Bootstrap 3.
* `lg` max width was 60rem which is 960px, less gutter than `md`.
  Changed to 940px, same as Bootstrap 3.
* `xl` max width was 72.25rem which is 1140px not 1156px.
  Changed to 1140px matching the comment but not the docs.

All the values are now the same as they were in Bootstrap 3.
Addresses #17070 and #17388.